move all inventory endpoints from BFF to InventoryControllerV1
