--- api-gateway/src/main/java/com/petclinic/bffapigateway/presentationlayer/v1/InventoryController.java
@@ -1,61 +1,0 @@
-package com.petclinic.bffapigateway.presentationlayer.v1;
-
-import com.petclinic.bffapigateway.domainclientlayer.InventoryServiceClient;
-import com.petclinic.bffapigateway.dtos.Inventory.*;
-import com.petclinic.bffapigateway.utils.Security.Annotations.SecuredEndpoint;
-import com.petclinic.bffapigateway.utils.Security.Variables.Roles;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-import reactor.core.publisher.Flux;
-import reactor.core.publisher.Mono;
-
-@RestController
-@RequiredArgsConstructor
-@Slf4j
-@RequestMapping("api/gateway/inventory")
-public class InventoryController {
-
-    private final InventoryServiceClient inventoryServiceClient;
-
-
-    @SecuredEndpoint(allowedRoles = {Roles.ADMIN,Roles.INVENTORY_MANAGER})
-    @PostMapping(value = "inventory",consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
-    public Mono<ResponseEntity<InventoryResponseDTO>> addInventory(@RequestBody InventoryRequestDTO model){
-        return inventoryServiceClient.addInventory(model)
-                .map(s -> ResponseEntity.status(HttpStatus.CREATED).body(s))
-                .defaultIfEmpty(ResponseEntity.badRequest().build());
-
-    }
-
-    @SecuredEndpoint(allowedRoles = {Roles.ADMIN,Roles.INVENTORY_MANAGER})
-    @PostMapping(value = "inventory/type")
-    public Mono<ResponseEntity<InventoryTypeResponseDTO>> addInventoryType(@RequestBody InventoryTypeRequestDTO inventoryTypeRequestDTO){
-        return inventoryServiceClient.addInventoryType(inventoryTypeRequestDTO)
-                .map(s -> ResponseEntity.status(HttpStatus.CREATED).body(s))
-                .defaultIfEmpty(ResponseEntity.badRequest().build());
-    }
-
-    @SecuredEndpoint(allowedRoles = {Roles.ADMIN, Roles.INVENTORY_MANAGER})
-    @PostMapping("/{inventoryId}/products")
-    public Mono<ResponseEntity<ProductResponseDTO>> addSupplyToInventory(
-            @RequestBody ProductRequestDTO productRequestDTO,
-            @PathVariable String inventoryId) {
-        return inventoryServiceClient.addSupplyToInventory(productRequestDTO, inventoryId)
-                .map(productResponseDTO -> ResponseEntity.status(HttpStatus.CREATED).body(productResponseDTO))
-                .defaultIfEmpty(ResponseEntity.badRequest().build());
-    }
-
-    @SecuredEndpoint(allowedRoles = {Roles.ADMIN,Roles.INVENTORY_MANAGER})
-    @GetMapping(value = "inventory/types")
-    public Flux<InventoryTypeResponseDTO> getAllInventoryTypes(){
-        return inventoryServiceClient.getAllInventoryTypes();
-    }
-
-
-
-
-}
